+++ R/FSelector.R
@@ -109,7 +109,11 @@ FSelector = R6Class("FSelector",
     #' @return [data.table::data.table()].
     optimize = function(inst) {
       assert_multi_class(inst, c("FSelectInstanceSingleCrit", "FSelectInstanceMultiCrit"))
-      optimize_default(inst, self, private)
+      inst$.__enclos_env__$private$.context = ContextOptimization$new(instance = inst, optimizer = self)
+      call_back("on_optimization_begin", inst$callbacks, get_private(inst)$.context)
+      result = optimize_default(inst, self, private)
+      call_back("on_optimization_end", inst$callbacks, get_private(inst)$.context)
+      result
     }
   ),